split different comm method for mnist distributed training test=develop
